+++ src/lib/form-group/form-group.component.ts
@@ -12,14 +12,15 @@ export class FormGroupType extends AbstractGroupType<IGroupTypeOptions> {
   template: `
     <ng-content></ng-content>
     <ng-container *ngFor="let field of mwFieldType.options.model | keyvalue:orderAsGiven">
-      <mw-form-field *ngIf="renderTargets[field.key] == null"
-                     [mwFormGroup]="mwElement"
-                     [mwElement]="elements[field.key]"
-                     [mwFieldType]="field.value"
-                     [mwPath]="fieldPaths[field.key]"
-                     [mwIndex]="indexFromParent"
-                     [mwSlots]="mwSlots">
-      </mw-form-field>
+      <ng-container mwFormField
+                    *ngIf="renderTargets[field.key] == null"
+                    [mwFormGroup]="mwElement"
+                    [mwElement]="elements[field.key]"
+                    [mwFieldType]="field.value"
+                    [mwPath]="fieldPaths[field.key]"
+                    [mwIndex]="indexFromParent"
+                    [mwSlots]="mwSlots">
+      </ng-container>
     </ng-container>
   `,
 })